fix(tsd): token verification and file import error
introduced the old configs back in .properties file

ElixirAuthService now verifies the authority using a
provided public key when extracting the subject from
the token issued by elixirAAI.

The initial maxChunk was missing in the resumable
upload, and now we set them in both ega and file
services.

Now, after a upload the response we get is a single
resumable instead of an array. It was a mistake
made during the rewrite.

Set the body to be optional when making the patch
request to avoid 400 client errors. Because when we
send off the "end" chunk we don't send anything.

Simplified the logic in processChunk.

Save the resumable after a model change. This was
missed during the rewrite.

Also, `md5Sum` json property to match the model prop.