feat(crypt4gh): Sign JARs and make publications for Maven Central - Add generation of JavaDocs and sources - Add generation of "fat JAR" that includes all external dependencies. This allows Crypt4GH to be run as a stand-alone tool. - Sign JARs using a GPG key - Configure 3 publications:     1) MavenJava: includes everything (slim JAR, fat JAR, docs and sources. These will be signed)     2) library: includes only the slim JAR and lists all external dependencies in the POM-file     3) tool: includes the fat JAR and lists no dependencies in the POM-file - Configure 3 repositories:     1) GitHub Packages (named "fega-norway-crypt4gh")     2) MavenCentral  (so far this only works for Snapshots)     3) LocalStaging  (a local directory where all the files needed by Maven Central can be staged)
It is now possible to publish SNAPSHOTS to Maven Central via the new Central Portal (authenticating with a pre-generated token).
The build-file also includes a block to use JReleaser to publish files from the local staging area to Maven Central
via the new Central Portal publishing API, but this might not work properly yet.
However, it should be possible to run the JReleaser command-line tool to publish staged files (if a "jreleaser.yml" configuration file is present).